fix: include react-router-dom dependency
