+++ frontend/package.json
@@ -12,7 +12,8 @@
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4",
     "chart.js": "^4.4.1",
-    "react-chartjs-2": "^5.3.1"
+    "react-chartjs-2": "^5.3.1",
+    "react-router-dom": "6.22.3"
   },
   "scripts": {
     "start": "react-scripts start",
